feat(backend): add the dynamo.backend SDK behind a compat shim
Part 1 of a 3-PR stack (add SDK -> migrate consumers -> rename).

Introduce the public `dynamo.backend` SDK and the `dynamo.sample_engine`
reference package, moving the real implementation out of the private
`dynamo.common.backend`. To keep this PR a pure addition with zero consumer
churn, `dynamo.common.backend` stays in place as a thin compatibility shim:
each old submodule (`engine`, `worker`, `run`, `publisher`, ...) aliases to its
new home, and the `-m dynamo.common.backend.sample_main` entry point still runs.

- `dynamo.backend`: facade (BaseEngine/LLMEngine/RawEngine/DiffusionEngine,
  EngineConfig, LlmRegistration, Generate*/Raw* types, Worker/WorkerConfig,
  add_worker_args, build_worker_config, run, Context, DisaggregationMode);
  private `_engine`/`_worker`/`_run`; advanced helper submodules; py.typed.
- `dynamo.sample_engine`: LLM + diffusion reference engine (`python -m
  dynamo.sample_engine`).
- CI/CODEOWNERS: add the new paths (old shim paths retained).

The shim is removed and consumers migrated in part 2; terminology is renamed
in part 3.
