+++ CODEOWNERS
@@ -230,6 +230,7 @@
 /lib/backend-common/examples/                                                @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/                                               @ai-dynamo/dynamo-runtime-codeowners
 /tests/test_models_dir_flag.py                                               @ai-dynamo/dynamo-runtime-codeowners
+/components/src/dynamo/backend/                                              @ai-dynamo/dynamo-runtime-codeowners
 /tests/report_pytest_markers.py                                              @ai-dynamo/dynamo-runtime-codeowners
 /examples/backends/tritonserver/                                             @ai-dynamo/dynamo-runtime-codeowners
 /tests/serve/multimodal_profiles/                                            @ai-dynamo/dynamo-runtime-codeowners
@@ -238,6 +239,7 @@
 /components/src/dynamo/common/tests/                                         @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/utils/                                         @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/memory/                                        @ai-dynamo/dynamo-runtime-codeowners
+/components/src/dynamo/sample_engine/                                        @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/backend/                                       @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/config_dump/                                   @ai-dynamo/dynamo-runtime-codeowners
 /components/src/dynamo/common/configuration/                                 @ai-dynamo/dynamo-runtime-codeowners
