+++ components/src/dynamo/backend/_run.py
@@ -1,0 +1,52 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Common entry point for backend workers.
+
+A backend's ``main`` calls :func:`run` with its ``BaseEngine`` subclass.
+Example::
+
+    from dynamo.backend import run
+    from dynamo.sample_engine import SampleLLMEngine
+
+    def main():
+        run(SampleLLMEngine)
+"""
+
+from collections.abc import Awaitable, Callable
+
+import uvloop
+
+from ._engine import BaseEngine
+from ._worker import Worker, WorkerConfig
+
+EngineFactory = Callable[[list[str] | None], Awaitable[tuple[BaseEngine, WorkerConfig]]]
+
+
+async def _start(
+    engine_cls: type[BaseEngine],
+    argv: list[str] | None = None,
+    engine_factory: EngineFactory | None = None,
+):
+    from_args = engine_factory or engine_cls.from_args
+    engine, worker_config = await from_args(argv)
+    w = Worker(engine, worker_config)
+    await w.run()
+
+
+def run(
+    engine_cls: type[BaseEngine],
+    argv: list[str] | None = None,
+    engine_factory: EngineFactory | None = None,
+):
+    """Entry point for per-backend ``<backend>_main.py`` files.
+
+    ``engine_cls`` may be an :class:`LLMEngine` or a :class:`DiffusionEngine`
+    subclass; both share the ``from_args -> (engine, WorkerConfig)`` contract.
+
+    ``engine_factory`` lets an entry point that already parsed args construct
+    its engine without parsing again. The default remains
+    ``engine_cls.from_args(argv)`` so backend-specific arguments do not leak
+    into the shared :class:`BaseEngine` contract.
+    """
+    uvloop.run(_start(engine_cls, argv, engine_factory))
